refactor: delete monitor http dependency shell
